fix: fixed app crash issue
Fixed app crash when selecting a category. Added pendingIntent implementation to delete a notification from map. The notification posted on selecting a category is posted on behalf of blurb so package name will no longer be the original application that posted the notification. However the key remains to be the same. Fixed notification classification on notifications posted by blurb.